Refactored (base) classes from init.py into separate files
